src/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ML and Search
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2

# LLM APIs
google-generativeai==0.3.1
requests==2.31.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.3
